Add FEntityId and FEntityIdManager
As fbx nodes have their own unique ids. We need to generate an id, which
will be unique for any entity when either import a scene or  create it
from scratch. If an entity is created from an fbx scene, its id is
stayed without touching. When create a custom entity, last id is used
and then it`s decremented. The initial value is max value of uint64_t